# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/orderitems": {
            "get": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Consultar todos los items de orden",
                "description": "Retorna todos los items de orden",
                "operationId": "57d67e1dd35a19d47dd8178eb254621e",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No se encontraron items"
                    },
                    "405": {
                        "description": "Método no permitido"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Crear item de orden",
                "description": "Crear un nuevo item de orden",
                "operationId": "32eae6df28d76229deca5bc0ed820210",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "product_id",
                                    "cantidad",
                                    "precio_unitario"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "precio_unitario": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item creado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/orderitems/{orderitem}": {
            "get": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Consultar un item de orden",
                "description": "Retorna un item de orden específico",
                "operationId": "4829fb193d059222e610aafa9249a8d7",
                "parameters": [
                    {
                        "name": "orderitem",
                        "in": "path",
                        "description": "ID del item de orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Actualizar item de orden",
                "description": "Actualizar un item de orden existente",
                "operationId": "00481fbca576cc3fffebc2e9b582b059",
                "parameters": [
                    {
                        "name": "orderitem",
                        "in": "path",
                        "description": "ID del item de orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cantidad": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "precio_unitario": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Item actualizado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OrderItems"
                ],
                "summary": "Eliminar item de orden",
                "description": "Elimina un item de orden existente",
                "operationId": "186caee0c62728042869ce6c5ab37506",
                "parameters": [
                    {
                        "name": "orderitem",
                        "in": "path",
                        "description": "ID del item de orden",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item eliminado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Item no encontrado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}